Clear up issues with .zshrc, link.sh, and my zsh theme

+++ setup.sh
@@ -5,8 +5,12 @@ SCRIPT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 
 for FILE in $DOTFILES; do
-	rm -f $HOME/$FILE
+	rm -rf $HOME/$FILE
 	ln -s $SCRIPT_DIR/$FILE $HOME/$FILE
 done
 
-ln -s $HOME/.vimrc $HOME/.nvimrc
+if [ ! -e $HOME/.nvimrc ]; then
+	ln -s $HOME/.vimrc $HOME/.nvimrc
+fi
+
+vim +PluginInstall +qall
